Rename api.c/h to <dir>.c/h
