+++ src/ckzg.c
@@ -20,9 +20,9 @@
 #include "common/fr.c"
 #include "common/lincomb.c"
 #include "common/utils.c"
-#include "eip4844/api.c"
 #include "eip4844/blob.c"
-#include "eip7594/api.c"
+#include "eip4844/eip4844.c"
 #include "eip7594/cell.c"
+#include "eip7594/eip7594.c"
 #include "eip7594/fft.c"
-#include "setup/api.c"
+#include "setup/setup.c"
